refactor: restructure vendor implementations into logical modules
Convert monolithic _implementation.py files into organized implementation/
subdirectories with logical separation of concerns:

- enums.py: Device-specific enumerations using IntEnum
- fields.py: BitField classes for low-level device control
- commands.py: Command construction classes (where applicable)
- state.py: Complete device state management
- __init__.py: Re-exports for backward compatibility

Affected vendors:
- ThingM: 197 lines → enums/fields/state modules
- Embrava: 146 lines → enums/fields/state modules
- EPOS: 78 lines → enums/fields/state modules
- Agile Innovative: 99 lines → state module
- MuteMe: 72 lines → fields/state modules
- Luxafor: 72 lines → enums/state modules
- Kuando: Already restructured as proof of concept

Benefits:
- Improved API documentation organization
- Better code discoverability and maintainability
- Logical separation of device control concerns
- Maintained backward compatibility through re-exports

All 721+ tests pass. Ruff formatting applied to ensure code quality.
