+++ src/busylight_core/vendors/agile_innovative/blinkstick_square.py
@@ -8,8 +8,8 @@ from typing import TYPE_CHECKING, ClassVar
 if TYPE_CHECKING:
     from busylight_core.hardware import Hardware
 
-from ._blinkstick import State
 from .blinkstick_base import BlinkStickBase
+from .implementation import State
 
 
 class BlinkStickSquare(BlinkStickBase):
